+++ src/app/child-dev-project/attendance/activity-attendance-section/activity-attendance-section.component.ts
@@ -8,7 +8,7 @@ import { AttendanceService } from "../attendance.service";
 import { PercentPipe } from "@angular/common";
 import { ActivityAttendance } from "../model/activity-attendance";
 import { Note } from "../../notes/model/note";
-import { ComponentWithConfig } from "../../../core/entity-components/entity-subrecord/entity-subrecord.component";
+import { ComponentWithConfig } from "../../../core/entity-components/entity-subrecord/entity-subrecord/entity-subrecord.component";
 import moment from "moment";
 
 @Component({
