+++ src/app/core/entity-components/entity-subrecord/entity-subrecord/entity-subrecord.component.html
