+++ src/app/core/entity-components/entity-subrecord/entity-subrecord/entity-subrecord.component.ts
@@ -13,18 +13,18 @@ import { MatSnackBar } from "@angular/material/snack-bar";
 import { MatSort } from "@angular/material/sort";
 import { MatPaginator, PageEvent } from "@angular/material/paginator";
 import { MatTableDataSource } from "@angular/material/table";
-import { ColumnDescription } from "./column-description";
+import { ColumnDescription } from "../column-description";
 import { MediaChange, MediaObserver } from "@angular/flex-layout";
-import { FormValidationResult } from "./form-validation-result";
-import { ColumnDescriptionInputType } from "./column-description-input-type.enum";
+import { FormValidationResult } from "../form-validation-result";
+import { ColumnDescriptionInputType } from "../column-description-input-type.enum";
 import { ComponentType } from "@angular/cdk/overlay";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
-import { EntityMapperService } from "../../entity/entity-mapper.service";
-import { Entity } from "../../entity/entity";
-import { ShowsEntity } from "../../form-dialog/shows-entity.interface";
-import { FormDialogService } from "../../form-dialog/form-dialog.service";
-import { ConfirmationDialogService } from "../../confirmation-dialog/confirmation-dialog.service";
-import { AlertService } from "../../alerts/alert.service";
+import { EntityMapperService } from "../../../entity/entity-mapper.service";
+import { Entity } from "../../../entity/entity";
+import { ShowsEntity } from "../../../form-dialog/shows-entity.interface";
+import { FormDialogService } from "../../../form-dialog/form-dialog.service";
+import { ConfirmationDialogService } from "../../../confirmation-dialog/confirmation-dialog.service";
+import { AlertService } from "../../../alerts/alert.service";
 import { DatePipe } from "@angular/common";
 import { BehaviorSubject } from "rxjs";
